update to .NET 7.0 and add processing time

+++ DotNetCoreSqlDb/Controllers/TodosController.cs
@@ -5,11 +5,14 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using DotNetCoreSqlDb;
+using DotNetCoreSqlDb.Data;
 using DotNetCoreSqlDb.Models;
 using Microsoft.Extensions.Caching.Distributed;
 
 namespace DotNetCoreSqlDb.Controllers
 {
+    [ActionTimerFilter]
     public class TodosController : Controller
     {
         private readonly MyDatabaseContext _context;
@@ -26,29 +29,18 @@ namespace DotNetCoreSqlDb.Controllers
         public async Task<IActionResult> Index()
         {
             var todos = new List<Todo>();
-            byte[] TodoListByteArray;
+            byte[]? TodoListByteArray;
 
-            // This allows the home page to load if migrations have not been run yet.
-            try
-            {
-                TodoListByteArray = await _cache.GetAsync(_TodoItemsCacheKey);
-                if (TodoListByteArray != null && TodoListByteArray.Length > 0)
-                { 
-                    todos = ConvertData<Todo>.ByteArrayToObjectList(TodoListByteArray);
-                }
-                else 
-                {
-                    todos = await _context.Todo.ToListAsync();
-                    TodoListByteArray = ConvertData<Todo>.ObjectListToByteArray(todos);
-                    await _cache.SetAsync(_TodoItemsCacheKey, TodoListByteArray);
-                }
-
-                
+            TodoListByteArray = await _cache.GetAsync(_TodoItemsCacheKey);
+            if (TodoListByteArray != null && TodoListByteArray.Length > 0)
+            { 
+                todos = ConvertData<Todo>.ByteArrayToObjectList(TodoListByteArray);
             }
-            catch (Exception e)
+            else 
             {
-
-                return View(todos);
+                todos = await _context.Todo.ToListAsync();
+                TodoListByteArray = ConvertData<Todo>.ObjectListToByteArray(todos);
+                await _cache.SetAsync(_TodoItemsCacheKey, TodoListByteArray);
             }
 
             return View(todos);
@@ -57,8 +49,8 @@ namespace DotNetCoreSqlDb.Controllers
         // GET: Todos/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            byte[] todoItemByteArray;
-            Todo todo;
+            byte[]? todoItemByteArray;
+            Todo? todo;
 
             if (id == null)
             {
@@ -75,10 +67,10 @@ namespace DotNetCoreSqlDb.Controllers
             {
                 todo = await _context.Todo
                 .FirstOrDefaultAsync(m => m.ID == id);
-                if (todo == null)
-                {
-                    return NotFound();
-                }
+            if (todo == null)
+            {
+                return NotFound();
+            }
 
                 todoItemByteArray = ConvertData<Todo>.ObjectToByteArray(todo);
                 await _cache.SetAsync(GetTodoItemCacheKey(id), todoItemByteArray);
@@ -96,8 +88,8 @@ namespace DotNetCoreSqlDb.Controllers
         }
 
         // POST: Todos/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("ID,Description,CreatedDate")] Todo todo)
@@ -129,8 +121,8 @@ namespace DotNetCoreSqlDb.Controllers
         }
 
         // POST: Todos/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("ID,Description,CreatedDate")] Todo todo)
@@ -189,10 +181,13 @@ namespace DotNetCoreSqlDb.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var todo = await _context.Todo.FindAsync(id);
-            _context.Todo.Remove(todo);
-            await _context.SaveChangesAsync();
-            await _cache.RemoveAsync(GetTodoItemCacheKey(todo.ID));
-            await _cache.RemoveAsync(_TodoItemsCacheKey);
+            if (todo != null)
+            {
+                _context.Todo.Remove(todo);
+                await _context.SaveChangesAsync();
+                await _cache.RemoveAsync(GetTodoItemCacheKey(todo.ID));
+                await _cache.RemoveAsync(_TodoItemsCacheKey);
+            }
             return RedirectToAction(nameof(Index));
         }
 
@@ -206,4 +201,6 @@ namespace DotNetCoreSqlDb.Controllers
             return _TodoItemsCacheKey+"_&_"+id;
         }
     }
+
+    
 }
